update0.9

+++ salesManagement/main/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import redirect
 from .forms import productform, batchform, salesform
 from django.contrib import messages
 from .models import commission, batch, currency,sales,product
+from django.conf import settings
+
 # Create your views here.
 
 
@@ -13,9 +15,13 @@ def addpro(request):
             proObj=product()
             proObj.product_type=form.cleaned_data['product_type']
             proObj.name=form.cleaned_data['name']
-            proObj.image=request.FILES['image']
+            if(len(request.FILES)>0):
+                proObj.image=request.FILES['image',False]
+                proObj.rimage="<img height='20%' width='100%' src='"+request.build_absolute_uri('/')+"media/ProductsImages/"+str(proObj.image)+"'/>"
+            else:
+                proObj.image=None
+                proObj.rimage=''
             proObj.description=form.cleaned_data['description']
-            print(proObj.image)
             proObj.save()
             messages.info(request, "Product added successfully!")
         else:
@@ -29,16 +35,21 @@ def rmit(request):
         if form.is_valid():
             salesObj=sales()
             salesObj.product_type=form.cleaned_data['product_type']
-            print(salesObj.product_type)
             salesObj.quant=form.cleaned_data['quant']
             salesObj.saleprice=form.cleaned_data['saleprice']
             salesObj.batchid=form.cleaned_data['batchid']
             salesObj.unitprofit=form.cleaned_data['batchid']
             batchObj = batch.objects.get(batchid__exact=salesObj.batchid)
+            if(salesObj.product_type != getattr(batchObj, 'product_type')):
+                messages.error(request, "The product is not in this batch, please choose the correct batch!")
+                return render(request, 'removeitem.html', {'form': salesform})
             batchprice = getattr(batchObj, 'unit_price')
             batchQuant = getattr(batchObj, 'quant')
             if batchQuant > salesObj.quant:
-                batch.objects.filter(batchid__exact=salesObj.batchid).update(quant=batchQuant-salesObj.quant)
+                newquant=batchQuant-salesObj.quant
+                batch.objects.filter(batchid__exact=salesObj.batchid).update(quant=newquant)
+                batch.objects.filter(batchid__exact=salesObj.batchid).update(total_cost=getattr(batchObj, 'unit_price')*newquant)
+
             elif batchQuant == salesObj.quant:
                 batch.objects.filter(batchid__exact=salesObj.batchid).delete()
                 salesObj.batchid=None
@@ -69,6 +80,7 @@ def addit(request):
             product_type = getattr(getattr(getattr(batchObj, 'product_type'), 'product_type'), 'product_type')
             exrate = getattr(currency.objects.get(name__exact=batchObj.currency), 'exrate')
             batchObj.unit_price *= exrate
+            batchObj.total_cost=form.cleaned_data['total_cost']*exrate
             add = getattr(commission.objects.get(product_type__exact=product_type), 'addfee')
             multiply = getattr(commission.objects.get(product_type__exact=product_type), 'multiplyfee')
             batchObj.minselling = (add+float(batchObj.unit_price))/(1-multiply)
@@ -98,3 +110,4 @@ def home(request):
 
 def login(request):
     return redirect('/accounts/login/')
+
